split conda_env and get_conda_env
